fix the librairy MCC

+++ asserv_position_timer/MCC.cpp
@@ -3,7 +3,7 @@
 
 MCC::MCC(uint8_t noMoteur, uint8_t freq) : _moteur(noMoteur, freq) {}
 
-void MCC::bouger(uint8_t vitesse) {
+void MCC::bouger(int vitesse) {
   if (vitesse == 0)
     _moteur.run(RELEASE);
   if (vitesse > 0) {
