+++ asserv_position_timer/MCC.h
@@ -6,7 +6,7 @@
 class MCC{    
   public :
     MCC(uint8_t noMoteur, uint8_t freq = DC_MOTOR_PWM_RATE);
-    void bouger(uint8_t vitesse);
+    void bouger(int vitesse);
 
   private :
     AF_DCMotor _moteur;
